Adds static scheduler and basic user programs printing information via UART.

+++ scripts/gen-offsets.sh
@@ -1,8 +1,10 @@
-#/bin/bash
+#/bin/sh
 
-TMP=$(mktemp)
+[ -z "$1" ] && exit 1
+[ -z "$2" ] && exit 2
+[ -z "$(command -v $CC)" ] && exit 3
 
-$CC -Iinc -S offsets.c -o $TMP
-grep "#define" $TMP | sed -e "s/^\s*//" > inc/offsets.h
-
+TMP=$(mktemp /tmp/XXXXXX.s)
+$CC -Iinc -S $1 -o $TMP
+grep "#define" $TMP | sed -e "s/^\s*//" > $2
 rm $TMP
